pyproject.toml
[tool.poetry]
name = "befitbackend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.97.0"
strawberry-graphql = {extras = ["fastapi"], version = "^0.185.0"}
uvicorn = "^0.22.0"
asyncpg = "^0.27.0"
alembic = "^1.11.1"
sqlalchemy = "^2.0.16"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
pytest = "^7.3.2"
pylint = "^2.17.4"
faker = "^18.10.1"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
